Moves HOC into app context
