+++ src/app/withAppContext.tsx
@@ -1,10 +1,10 @@
-import React from 'react'
+import React, { ReactElement } from 'react'
 import { Provider } from 'react-redux'
 import { Router } from 'react-router-dom'
 import { ThemeProvider } from '@amsterdam/asc-ui'
 import '@testing-library/jest-dom'
 
-import configureStore from '../src/store/store'
+import configureStore from '../store/store'
 
 const { store, history } = configureStore()
 
@@ -12,10 +12,10 @@ const { store, history } = configureStore()
  * HOC that wraps a component in the necessary providers so that the component can access theming,
  * the store and can use routing.
  */
-const withAppContext = (Component) => (
+const withAppContext = (WrappedComponent: ReactElement) => (
   <ThemeProvider>
     <Provider store={store}>
-      <Router history={history}>{Component}</Router>
+      <Router history={history}>{WrappedComponent}</Router>
     </Provider>
   </ThemeProvider>
 )
